Добавил возможность добавлять блюдо в меню

+++ MenuWF/Entities/Journal.cs
@@ -3,7 +3,7 @@
 public class Journal : BaseEntity
 {
     public int RecipeId { get; set; }
-    public int DishWeight { get; set; }
+    public decimal DishWeight { get; set; }
     public DateTime Date { get; set; }
     public Meal meal { get; set; }
